# docs/create-class/openapi.yaml
openapi: 3.0.3
info:
  title: TMS - Class Creation Workflow API
  description: |
    # Class Creation Workflow API Documentation

    This API specification covers the complete **7-step workflow** for creating and managing classes
    in the Trainning Management System (TMS).

    ## Workflow Steps

    1. **Create Class** - Academic Staff creates a new class with basic information
    2. **Generate Sessions** - System automatically generates sessions based on course template
    3. **Assign Time Slots** - Academic Staff assigns time slots for each schedule day
    4. **Assign Resources** - Academic Staff assigns rooms/resources with auto-propagation
    5. **Assign Teachers** - Academic Staff assigns teachers with PRE-CHECK availability
    6. **Validate** - System validates completeness before submission
    7. **Submit for Approval** - Academic Staff submits, Center Head approves/rejects

    ## Key Features

    - **PRE-CHECK Approach** (v1.1): Teachers queried WITH availability status upfront
    - **Smart Matching**: Students matched by skill assessment
    - **HYBRID Auto-propagation**: SQL bulk operations + detailed conflict analysis
    - **'general' Skill**: Universal skill that can teach any session

    ## Authentication

    All endpoints require JWT Bearer authentication with role-based authorization:
    - `ACADEMIC_STAFF`: Steps 1-7 (create, assign, submit)
    - `CENTER_HEAD`: Approval/rejection

  version: 1.1.0
  contact:
    name: TMS Development Team
...
servers:
...
paths:
...
  /classes:
    post:
      tags:
        - Step 1 - Create Class
      summary: Create a new class
      description: |
        **STEP 1** of the workflow: Academic Staff creates a new class.

        **Validations:**
        - `start_date` must be in `schedule_days`
        - Course must be approved and active
        - `(branch_id, code)` must be unique

        **After creation**, the system automatically triggers **STEP 2** (Generate Sessions).

      operationId: createClass
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClassRequest'
            examples:
              offline_class:
                summary: Offline English A1 class
                value:
                  branchId: 1
                  courseId: 10
                  code: "ENG-A1-2024-01"
                  name: "English A1 Foundation - Morning"
                  modality: "OFFLINE"
                  startDate: "2025-01-06"
                  scheduleDays: [1, 3, 5]
                  maxCapacity: 20
              online_class:
                summary: Online IELTS class
                value:
                  branchId: 2
                  courseId: 15
                  code: "IELTS-ADV-2024-02"
                  name: "IELTS Advanced - Evening"
                  modality: "ONLINE"
                  startDate: "2025-01-08"
                  scheduleDays: [2, 4]
                  maxCapacity: 25
      responses:
        '201':
          description: Class created successfully, sessions auto-generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateClassResponse'
              examples:
                success:
                  summary: Class created with sessions
                  value:
                    success: true
                    message: "Class created successfully with 36 sessions generated"
                    data:
                      classId: 101
                      code: "ENG-A1-2024-01"
                      branchId: 1
                      courseId: 10
                      modality: "OFFLINE"
                      startDate: "2025-01-06"
                      scheduleDays: [1, 3, 5]
                      maxCapacity: 20
                      status: "DRAFT"
                      approvalStatus: "PENDING"
                      sessionsGenerated: 36
                      createdAt: "2025-01-03T10:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Conflict - Class code already exists for this branch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "Class code 'ENG-A1-2024-01' already exists for branch 'Main Campus'"
                errorCode: "CLASS_CODE_DUPLICATE"
                timestamp: "2025-01-03T10:30:00Z"

    get:
      tags:
        - Utilities
      summary: List classes with filters
      description: Get paginated list of classes with filtering and search capabilities
      operationId: getClasses
      security:
        - bearerAuth: []
      parameters:
        - name: branchIds
          in: query
          description: Filter by branch ID(s)
          schema:
            type: array
            items:
              type: integer
              format: int64
        - name: courseId
          in: query
          description: Filter by course ID
          schema:
            type: integer
            format: int64
        - name: modality
          in: query
          description: Filter by modality
          schema:
            $ref: '#/components/schemas/Modality'
        - name: search
          in: query
          description: Search in class code, name, course name, branch name
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          schema:
            type: string
            default: "startDate"
        - name: sortDir
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: "asc"
      responses:
        '200':
          description: Classes retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassListResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/v1/auth/login
...
  schemas:
...
    CreateClassRequest:
      type: object
      required:
        - branchId
        - courseId
        - code
        - modality
        - startDate
        - scheduleDays
        - maxCapacity
      properties:
        branchId:
          type: integer
          format: int64
          description: Branch ID where the class will be held
          example: 1
        courseId:
          type: integer
          format: int64
          description: Course template ID to use for session generation
          example: 10
        code:
          type: string
          maxLength: 50
          description: Unique class code (within branch)
          example: "ENG-A1-2024-01"
          pattern: '^[A-Z0-9-]+$'
        name:
          type: string
          maxLength: 255
          description: Optional descriptive name for the class
          example: "English A1 Foundation - Morning"
        modality:
          $ref: '#/components/schemas/Modality'
        startDate:
          type: string
          format: date
          description: Class start date (must be in scheduleDays)
          example: "2025-01-06"
        scheduleDays:
          type: array
          description: Days of week for class sessions (PostgreSQL DOW - 0=Sun, 1=Mon, ..., 6=Sat)
          items:
            type: integer
            minimum: 1
            maximum: 7
          minItems: 1
          maxItems: 7
          example: [1, 3, 5]
        maxCapacity:
          type: integer
          minimum: 1
          description: Maximum number of students allowed
          example: 20
    CreateClassResponse:
      allOf:
        - $ref: '#/components/schemas/ResponseWrapper'
        - type: object
          properties:
            data:
              type: object
              properties:
                classId:
                  type: integer
                  format: int64
                  example: 101
                code:
                  type: string
                  example: "ENG-A1-2024-01"
                branchId:
                  type: integer
                  format: int64
                courseId:
                  type: integer
                  format: int64
                modality:
                  $ref: '#/components/schemas/Modality'
                startDate:
                  type: string
                  format: date
                scheduleDays:
                  type: array
                  items:
                    type: integer
                maxCapacity:
                  type: integer
                status:
                  $ref: '#/components/schemas/ClassStatus'
                approvalStatus:
                  $ref: '#/components/schemas/ApprovalStatus'
                sessionsGenerated:
                  type: integer
                  description: Number of sessions auto-generated
                  example: 36
                createdAt:
                  type: string
                  format: date-time

    # ========== STEP 3: ASSIGN TIME SLOTS SCHEMAS ==========
...
    ClassListResponse:
      allOf:
        - $ref: '#/components/schemas/ResponseWrapper'
        - type: object
          properties:
            data:
              type: object
              properties:
                content:
                  type: array
                  items:
                    type: object
                    properties:
                      classId:
                        type: integer
                        format: int64
                      code:
                        type: string
                      name:
                        type: string
                      branchName:
                        type: string
                      courseName:
                        type: string
                      modality:
                        $ref: '#/components/schemas/Modality'
                      startDate:
                        type: string
                        format: date
                      status:
                        $ref: '#/components/schemas/ClassStatus'
                      approvalStatus:
                        $ref: '#/components/schemas/ApprovalStatus'
                      currentEnrollment:
                        type: integer
                      maxCapacity:
                        type: integer
                page:
                  type: integer
                size:
                  type: integer
                totalElements:
                  type: integer
                  format: int64
                totalPages:
                  type: integer
...
    Modality:
      type: string
      enum:
        - ONLINE
        - OFFLINE
        - HYBRID
      description: Class delivery modality
    ClassStatus:
      type: string
      enum:
        - DRAFT
        - SCHEDULED
        - ONGOING
        - COMPLETED
        - CANCELLED
      description: Current class status
    ApprovalStatus:
      type: string
      enum:
        - PENDING
        - APPROVED
        - REJECTED
      description: Approval status from Center Head
...
    ResponseWrapper:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        message:
          type: string
          description: Human-readable message
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ResponseWrapper'
        - type: object
          properties:
            errorCode:
              type: string
              description: Machine-readable error code
            details:
              type: object
              description: Additional error details
              additionalProperties: true
  responses:
    BadRequest:
      description: Bad request - Invalid input parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Invalid request parameters"
            errorCode: "BAD_REQUEST"
            timestamp: "2025-01-03T10:30:00Z"
    Unauthorized:
      description: Unauthorized - Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Authentication required"
            errorCode: "UNAUTHORIZED"
            timestamp: "2025-01-03T10:30:00Z"
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "You do not have permission to perform this action"
            errorCode: "FORBIDDEN"
            timestamp: "2025-01-03T10:30:00Z"
...
security:
  - bearerAuth: []